function/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-common==1.1.27
azure-core==1.13.0
azure-functions==1.7.0
azure-servicebus==7.1.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
psycopg2-binary==2.8.6
python-http-client==3.3.2
requests==2.25.1
requests-oauthlib==1.3.0
sendgrid==6.6.0
six==1.15.0
starkbank-ecdsa==1.1.0
uamqp==1.3.0
urllib3==1.26.4